requirement.txt
attrs==22.2.0
cmake==3.26.3
filelock==3.11.0
importlib-resources==5.12.0
Jinja2==3.1.2
JsonForm==0.0.2
jsonschema==4.17.3
JsonSir==0.0.2
lit==16.0.1
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
pygame==2.3.0
PyOpenGL==3.1.6
pyrsistent==0.19.3
Python-EasyConfig==0.1.7
PyYAML==6.0
Resource==0.2.1
six==1.16.0
sympy==1.11.1
torch==2.0.0
triton==2.0.0
typing_extensions==4.5.0
zipp==3.15.0
